EmbedAPI: make CDN cache responses to work
EmbedAPI v3 started failed when CDN was implemented on these views because the
`_get_project` and `_get_version` methods weren't implemented on this class.

This commit re-uses the `_get_project` and `_get_version` from the v2 version
and put it on a mixin that it's used by both APIs v2 and v3 to share the same code.
