+++ readthedocs/embed/views.py
@@ -1,13 +1,10 @@
 """Views for the embed app."""
 
-import functools
 import json
 import re
 
 import structlog
-from django.shortcuts import get_object_or_404
 from django.template.defaultfilters import slugify
-from django.utils.functional import cached_property
 from docutils.nodes import make_id
 from pyquery import PyQuery as PQ  # noqa
 from rest_framework import status
@@ -15,14 +12,12 @@ from rest_framework.renderers import BrowsableAPIRenderer, JSONRenderer
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from readthedocs.api.mixins import CDNCacheTagsMixin
+from readthedocs.api.mixins import CDNCacheTagsMixin, EmbedAPIMixin
 from readthedocs.api.v2.permissions import IsAuthorizedToViewVersion
 from readthedocs.builds.constants import EXTERNAL
 from readthedocs.core.resolver import resolve
-from readthedocs.core.unresolver import unresolve
 from readthedocs.core.utils.extend import SettingsOverrideObject
 from readthedocs.embed.utils import clean_links, recurse_while_none
-from readthedocs.projects.models import Project
 from readthedocs.storage import build_media_storage
 
 log = structlog.get_logger(__name__)
@@ -35,7 +30,7 @@ def escape_selector(selector):
     return ret
 
 
-class EmbedAPIBase(CDNCacheTagsMixin, APIView):
+class EmbedAPIBase(EmbedAPIMixin, CDNCacheTagsMixin, APIView):
 
     # pylint: disable=line-too-long
 
@@ -67,30 +62,6 @@ class EmbedAPIBase(CDNCacheTagsMixin, APIView):
 
     permission_classes = [IsAuthorizedToViewVersion]
     renderer_classes = [JSONRenderer, BrowsableAPIRenderer]
-
-    @functools.lru_cache(maxsize=1)
-    def _get_project(self):
-        if self.unresolved_url:
-            project_slug = self.unresolved_url.project.slug
-        else:
-            project_slug = self.request.GET.get('project')
-        return get_object_or_404(Project, slug=project_slug)
-
-    @functools.lru_cache(maxsize=1)
-    def _get_version(self):
-        if self.unresolved_url:
-            version_slug = self.unresolved_url.version_slug
-        else:
-            version_slug = self.request.GET.get('version', 'latest')
-        project = self._get_project()
-        return get_object_or_404(project.versions.all(), slug=version_slug)
-
-    @cached_property
-    def unresolved_url(self):
-        url = self.request.GET.get('url')
-        if not url:
-            return None
-        return unresolve(url)
 
     def get(self, request):
         """Handle the get request."""
